pyproject.toml
[project]
name = "fastapi-identity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nshathish", email = "470487+nshathish@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.37",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
